rename folders test data for exclusion set problem
