+++ README.md
@@ -37,8 +37,8 @@ Time complexity of algorithm in BigO is `O(|U| * log|S|)`.
 - Feature: Check a second set for elements to be covered or avoided
 - Feature: Priority queue 
 - Feature: Improved weighted set data structures
-- CI/CD
-- Lock master, feature branches only
+- CI/CD, pre-commit hooks (flake8, black)
+- Lock master to unapproved commits
 
 ## References:
