+++ requirements.txt
@@ -2,3 +2,4 @@
 pytest
 pandas
 setuptools
+tqdm
